Fixup build script to handle dependencies better; Add base for Linux and PowerPC support.

+++ build.sh
@@ -22,30 +22,152 @@ then
 	# Build hairball
 	if ! "$0" "build" "hairball"
 	then
-		echo "Failed to build hairball."
+		echo "Failed to build hairball." 2>&1
 		exit 1
 	fi
 	
 	# Launch hairball
-	"$0" "launch" $*
+	"$0" "launch" "hairball"
 	exit $?
 fi
 
 # Launch hairball
 if [ "$1" = "launch" ]
 then
-	java -cp "$(__i=0; for __file in *.jar; do if [ "$__i" -ne "0" ];
-		then echo -n ":"; fi; __i=$(($__i + 1)); echo -n $__file; done)" \
+	# Need arguments?
+	if [ "$#" -le "1" ]
+	then
+		echo "Usage: $0 $1 (package)" 2>&1
+		exit 1
+	fi
+	
+	# The package to launch
+	__pack="$2"
+	
+	# Launch the given command
+	java -cp "$("$0" "launchclasspath" "$__pack")" \
 		net.multiphasicapps.hairball.Main $*
 	exit $?
 
+# Calculate dependenices for a package to launch it
+elif [ "$1" = "launchclasspath" ]
+then
+	# Need arguments?
+	if [ "$#" -le "1" ]
+	then
+		echo "Usage: $0 $1 (package)" 2>&1
+		exit 1
+	fi
+	
+	# The package to get dependencies for
+	__pack="$2"
+	
+	# Get the dependent class path
+	__dcp="$("$0" dependsclasspath "$__pack")"
+	
+	# If there are none, then just print this package
+	if [ -z "$__dcp" ]
+	then
+		echo "$__pack.jar"
+	
+	# Otherwise add ours on top
+	else
+		echo "$__dcp:$__pack.jar"
+	fi
+
+# Calculate the dependent classpath for a specific package
+elif [ "$1" = "dependsclasspath" ]
+then
+	# Need arguments?
+	if [ "$#" -le "1" ]
+	then
+		echo "Usage: $0 $1 (package)" 2>&1
+		exit 1
+	fi
+	
+	# The package to get dependencies for
+	__pack="$2"
+	
+	# Get dependencies
+	__i=0
+	"$0" depends "$__pack" | while read __line
+	do
+		# Flagged?
+		if [ "$__i" -ne "0" ]
+		then
+			echo -n ":"
+		fi
+		
+		# Set to use colon next time
+		__i=1
+		
+		# Print it
+		echo -n "$__line.jar"
+	done
+
+# Get all package dependencies without recursion
+elif [ "$1" = "dependsnorecurse" ]
+then
+	# Need arguments?
+	if [ "$#" -le "1" ]
+	then
+		echo "Usage: $0 $1 (package)" 2>&1
+		exit 1
+	fi
+	
+	# The package to get dependencies for
+	__pack="$2"
+
+	# Needs a manifest
+	__manf="$__exedir/src/$__pack/META-INF/MANIFEST.MF"
+	if [ ! -f "$__manf" ]
+	then
+		echo "The package $__pack does not have a manifest." 2>&1
+		exit 1
+	fi
+	
+	# Echo them out
+	(for __dep in $(tr '\n' '\v' < $__manf | sed 's/\v //g' | tr '\v' '\n' |
+		grep '^X-Hairball-Depends[ \t]*:' | sed 's/^[^:]*://g')
+	do
+		# Print this dependency
+		echo "$__dep"
+	done) | sort | uniq
+
+# Get all package dependencies (recursive)
+elif [ "$1" = "depends" ]
+then
+	# Need arguments?
+	if [ "$#" -le "1" ]
+	then
+		echo "Usage: $0 $1 (package)" 2>&1
+		exit 1
+	fi
+	
+	# The package to get dependencies for
+	__pack="$2"
+	
+	# Call non-recursive getter
+	("$0" "dependsnorecurse" "$__pack" | while read __dep
+	do
+		# Print this dependency
+		echo "$__dep"
+		
+		# Print dependencies of that dependency
+		if ! "$0" "depends" "$__dep"
+		then
+			echo "Failed to print dependency $__dep for package $__pack."
+			exit 1
+		fi
+	done) | sort | uniq
+	
 # Building a package
 elif [ "$1" = "build" ]
 then
 	# Need arguments?
 	if [ "$#" -le "1" ]
 	then
-		echo "Usage: $0 $1 (package)"
+		echo "Usage: $0 $1 (package)" 2>&1
 		exit 1
 	fi
 	
@@ -56,19 +178,17 @@ then
 	__manf="$__exedir/src/$__pack/META-INF/MANIFEST.MF"
 	if [ ! -f "$__manf" ]
 	then
-		echo "The package $__pack does not have a manifest."
+		echo "The package $__pack does not have a manifest." 2>&1
 		exit 1
 	fi
 
 	# Build package dependencies
-	__depends="$(tr '\n' '\v' < $__manf | sed 's/\v //g' | tr '\v' '\n' |
-		grep '^X-Hairball-Depends[ \t]*:' | sed 's/^[^:]*://g')"
-	for __dep in $__depends
+	"$0" "dependsnorecurse" "$__pack" | while read __dep
 	do
 		# Build dependency
 		if ! "$0" "build" "$__dep"
 		then
-			echo "Failed to build dependency $__dep for $__pack."
+			echo "Failed to build dependency $__dep for $__pack." 2>&1
 			exit 1
 		fi
 	done
@@ -92,7 +212,7 @@ then
 		done
 	
 		# Compare dependencies (if they exist)
-		for __dep in $__depends
+		"$0" "dependsnorecurse" "$__pack" | while read __dep
 		do
 			if [ -f "$__dep.jar" ] && [ "$__dep.jar" -nt "$__ojar" ]
 			then
@@ -118,6 +238,9 @@ then
 	# Perform a rebuild?
 	if [ "$__ood" -ne "0" ]
 	then
+		# Delete the output JAR first
+		rm -f "$__ojar"
+		
 		# List source code which needs compilation
 		find "$__isrc" -type f | grep '\.java$' > /tmp/$$
 		
@@ -125,41 +248,33 @@ then
 		mkdir -p "/tmp/$$.$__pack"
 		
 		# Note it
-		echo "*** Compiling $__pack ***"
+		echo "*** Compiling $__pack ***" 2>&1
 		
 		# Calculate stuff to run it with
-		__cp=""
-		for __x in $__depends
-		do
-			# Add separator
-			if [ -n "$__cp" ]
-			then
-				__cp="$__cp:"
-			fi
-			
-			# Append
-			__cp="$__cp$__x.jar"
-		done
+		__cp="$("$0" "dependsclasspath" "$__pack")"
 		
-		# Compile code
-		if ! javac -d "/tmp/$$.$__pack" -source 1.7 -target 1.7 -g \
-			-bootclasspath "$__cp" @/tmp/$$
+		# Compile code but only if there is source code to actually compile
+		if [ -s "/tmp/$$" ]
 		then
-			# Note
-			echo "Failed to compile $__pack."
+			if ! javac -d "/tmp/$$.$__pack" -source 1.7 -target 1.7 -g \
+				-bootclasspath "$__cp" @/tmp/$$
+			then
+				# Note
+				echo "Failed to compile $__pack." 2>&1
 			
-			# Delete temporaries
-			rm -rf "/tmp/$$.$__pack"
+				# Delete temporaries
+				rm -rf "/tmp/$$.$__pack"
 			
-			# Failed
-			exit 1
+				# Failed
+				exit 1
+			fi
 		fi
 		
 		# Package it up
 		if ! jar cf "$__ojar" -C "/tmp/$$.$__pack" .
 		then
 			# Note
-			echo "Failed to package $__pack."
+			echo "Failed to package $__pack." 2>&1
 			
 			# Delete
 			rm -rf "/tmp/$$.$__pack"
@@ -174,7 +289,7 @@ then
 
 # Unknown command
 else
-	echo "Unknown sub-command $1."
+	echo "Unknown sub-command $1." 2>&1
 	exit 1
 fi
